+++ src/Exception/InvalidSortOperationException.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\RadBundle\Exception;
+
+class InvalidSortOperationException extends RadException
+{
+
+}